Improve dev ux for running tests

+++ docker/openmage/install_openmage.sh
@@ -2,6 +2,8 @@
 
 set -eu -o pipefail
 
+BASEDIR=$(dirname "${BASH_SOURCE:-$0}")
+
 if [ -z "${OPENMAGE_ROOT-}" ]; then
   echo "Environment variable OPENMAGE_ROOT not set"
   exit 1;
@@ -16,13 +18,11 @@ RELEASE_PATH="https://github.com/OpenMage/magento-lts/archive/refs/tags/%version
 
 echo -e "Using OpenMage Release \033[0;1;32m${OPENMAGE_VERSION}\033[0m"
 
-if [ -d "$OPENMAGE_ROOT" ]; then
-  rm -rf "$OPENMAGE_ROOT"
-fi
-
+[ -d "$OPENMAGE_ROOT" ] && rm -rf "$OPENMAGE_ROOT"
 mkdir -p "$OPENMAGE_ROOT"
 echo "${RELEASE_PATH/\%version\%/$OPENMAGE_VERSION}"
 curl -Ls "${RELEASE_PATH/\%version\%/$OPENMAGE_VERSION}" -o - | tar -C "$OPENMAGE_ROOT" -zx --strip-components 1
 /usr/bin/env composer -d "${OPENMAGE_ROOT}" install --ignore-platform-reqs
 
-cp patches/fix-zf1future-autoloader.php "${OPENMAGE_ROOT}/app/etc/includes/zf1future-autoload-fix.php"
+[ ! -d "${OPENMAGE_ROOT}/app/etc/includes/" ] && mkdir -p "${OPENMAGE_ROOT}/app/etc/includes/"
+cp "${BASEDIR}/patches/fix-zf1future-autoloader.php" "${OPENMAGE_ROOT}/app/etc/includes/zf1future-autoload-fix.php"
